Binance/requirements.txt
Package            Version    
------------------ -----------
python-binance     1.0.19     
